Add files via upload
It is the game of sinking the fleet, but very simplified, it is made up of an array of length 10, randomly creating ships "B" and water "A", the code is created in Spanish for that reason those names. The game gives 6 chances to guess the boats, of which 4 are boats and the rest is water.